Allow Journal.d on Ubuntu (jetson)

+++ stages/03-Preconfiguration/00-run-chroot.sh
@@ -46,7 +46,10 @@ sudo systemctl disable dhcpcd.service
 sudo systemctl disable dnsmasq.service
 sudo systemctl disable cron.service
 sudo systemctl disable syslog.service
-sudo systemctl disable journald.service
+if [[ "${OS}" != "ubuntu" ]]; then
+    echo "OS is NOT ubuntu..disabling journald"
+    sudo systemctl disable journald.service
+fi
 sudo systemctl disable triggerhappy.service
 sudo systemctl disable avahi-daemon.service
 sudo systemctl disable ser2net.service
@@ -61,15 +64,21 @@ sudo systemctl mask plymouth-start.service
 sudo systemctl mask plymouth-read-write.service
 sudo systemctl mask plymouth-quit-wait.service
 sudo systemctl mask plymouth-quit.service
-sudo systemctl disable systemd-journal-flush.service
+if [[ "${OS}" != "ubuntu" ]]; then
+    echo "OS is NOT ubuntu..disabling journald flush"
+    sudo systemctl disable systemd-journal-flush.service
+fi
 #this service updates runlevel changes. Set desired runlevel prior to this being disabled
 sudo systemctl disable systemd-update-utmp.service
 sudo systemctl disable networking.service
 
 #Mask difficult to disable services
-systemctl stop systemd-journald.service
-systemctl disable systemd-journald.service
-systemctl mask systemd-journald.service
+if [[ "${OS}" != "ubuntu" ]]; then
+    echo "OS is NOT ubuntu..stopping-disbale-mask journald"
+    systemctl stop systemd-journald.service
+    systemctl disable systemd-journald.service
+    systemctl mask systemd-journald.service
+fi
 
 
 sudo rm /etc/init.d/resize2fs_once
